# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TeleCloud",
    "description": "A cloud storage system that uses Telegram's Bot API as the storage backend.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/folders": {
      "get": {
        "tags": [
          "folders"
        ],
        "summary": "List Root",
        "description": "List the caller's root contents (``parent_id IS NULL``).",
        "operationId": "list_root_folders_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderContentsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "folders"
        ],
        "summary": "Create Folder",
        "description": "Create a folder (root, or under ``parent_id`` if given).",
        "operationId": "create_folder_folders_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFolderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/folders/{folder_id}": {
      "get": {
        "tags": [
          "folders"
        ],
        "summary": "List Folder",
        "description": "List one folder's subfolders and files.",
        "operationId": "list_folder_folders__folder_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "folder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Folder Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderContentsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "folders"
        ],
        "summary": "Rename Folder",
        "description": "Rename a folder.",
        "operationId": "rename_folder_folders__folder_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "folder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Folder Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameFolderRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "folders"
        ],
        "summary": "Delete Folder",
        "description": "Soft-delete a folder, cascading to its descendants and their files.\n\nFolder rows are marked ``deleted_at``; contained files are handed to\n``files/``'s deletion path (SPEC \u00a76.12). Telegram messages are NOT deleted\nhere \u2014 that is deferred to a ``jobs/`` deletion job (SPEC \u00a77.4).",
        "operationId": "delete_folder_folders__folder_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "folder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Folder Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/folders/{folder_id}/move": {
      "post": {
        "tags": [
          "folders"
        ],
        "summary": "Move Folder",
        "description": "Re-parent a folder (``new_parent_id: null`` moves it to the root).\n\nRejects cycles (a folder cannot become its own descendant) with\n``validation_error``.",
        "operationId": "move_folder_folders__folder_id__move_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "folder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Folder Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveFolderRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/shares": {
      "post": {
        "tags": [
          "sharing"
        ],
        "summary": "Create Share",
        "description": "Create a public share link for one of the caller's committed files.\n\nOptional ``expires_at`` (future, timezone-aware) and ``download_limit``\n(positive int) narrow the link; omitting them yields the defaults\n(never-expire, unlimited).",
        "operationId": "create_share_shares_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateShareRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "sharing"
        ],
        "summary": "List Shares",
        "description": "List every share link the caller created for one of their files (``?file_id=``).\n\nIncludes revoked links so a management UI can show the full history.",
        "operationId": "list_shares_shares_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "file_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "File Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/shares/{share_id}/revoke": {
      "post": {
        "tags": [
          "sharing"
        ],
        "summary": "Revoke Share",
        "description": "Revoke one of the caller's share links (soft: ``revoked=true``).\n\nReturns the updated row; revoking an already-revoked link is idempotent.\nSubsequent public downloads of the token are rejected with ``share_revoked``.",
        "operationId": "revoke_share_shares__share_id__revoke_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "share_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Share Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateFolderRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CreateFolderRequest",
        "description": "Body of ``POST /folders`` \u2014 create a folder, optionally under a parent."
      },
      "CreateShareRequest": {
        "properties": {
          "file_id": {
            "type": "string",
            "format": "uuid",
            "title": "File Id"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "download_limit": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Download Limit"
          }
        },
        "type": "object",
        "required": [
          "file_id"
        ],
        "title": "CreateShareRequest",
        "description": "Body of ``POST /shares`` \u2014 create a public link for one of the caller's files.\n\n``expires_at`` and ``download_limit`` are optional; omitting them yields the\nmodule defaults (never-expire, unlimited \u2014 see ``sharing/README.md``). When\ngiven they are validated in the service: ``expires_at`` must be timezone-aware\nand in the future, ``download_limit`` must be a positive integer."
      },
      "FileEntry": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "folder_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "mime_type": {
            "type": "string",
            "title": "Mime Type"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "folder_id",
          "name",
          "size_bytes",
          "mime_type",
          "created_at"
        ],
        "title": "FileEntry",
        "description": "A file as it appears in a folder listing (a narrow ``FileMeta`` view)."
      },
...
      "FolderContentsResponse": {
        "properties": {
          "folder_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder Id"
          },
          "folders": {
            "items": {
              "$ref": "#/components/schemas/FolderResponse"
            },
            "type": "array",
            "title": "Folders"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/FileEntry"
            },
            "type": "array",
            "title": "Files"
          }
        },
        "type": "object",
        "required": [
          "folder_id",
          "folders",
          "files"
        ],
        "title": "FolderContentsResponse",
        "description": "Contents of one folder: its child folders and its files.\n\n``folder_id`` is the listed folder (``null`` for the owner's root). Only\nnon-deleted subfolders and ``committed`` files are returned."
      },
      "FolderResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "parent_id",
          "name",
          "created_at"
        ],
        "title": "FolderResponse",
        "description": "Response-facing view of a folder row (no ``deleted_at``)."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MoveFolderRequest": {
        "properties": {
          "new_parent_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "New Parent Id"
          }
        },
        "type": "object",
        "title": "MoveFolderRequest",
        "description": "Body of ``POST /folders/{folder_id}/move`` \u2014 the new parent (``null`` = root)."
      },
...
      "RenameFolderRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "RenameFolderRequest",
        "description": "Body of ``PATCH /folders/{folder_id}`` \u2014 the new name."
      },
...
      "ShareListResponse": {
        "properties": {
          "file_id": {
            "type": "string",
            "format": "uuid",
            "title": "File Id"
          },
          "shares": {
            "items": {
              "$ref": "#/components/schemas/ShareResponse"
            },
            "type": "array",
            "title": "Shares"
          }
        },
        "type": "object",
        "required": [
          "file_id",
          "shares"
        ],
        "title": "ShareListResponse",
        "description": "The share links created for one of the caller's files (management view)."
      },
      "ShareResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "file_id": {
            "type": "string",
            "format": "uuid",
            "title": "File Id"
          },
          "token": {
            "type": "string",
            "title": "Token"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "download_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Download Limit"
          },
          "download_count": {
            "type": "integer",
            "title": "Download Count"
          },
          "revoked": {
            "type": "boolean",
            "title": "Revoked"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "file_id",
          "token",
          "expires_at",
          "download_limit",
          "download_count",
          "revoked",
          "created_at"
        ],
        "title": "ShareResponse",
        "description": "Response-facing view of a share row \u2014 owner identity intentionally omitted.\n\nMirrors the meaningful columns of ``public.shares`` (SPEC \u00a74.5) minus\n``owner_id``. The ``token`` is returned so the caller can construct the public\n``/s/{token}`` URL (the host is the caller's concern; ``sharing/`` does not know\nthe deployment origin)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}